+++ aries_vcx_core/src/indy/credentials/issuer/mod.rs
@@ -5,7 +5,7 @@ use crate::global::settings;
 use crate::indy::anoncreds;
 use crate::utils::constants::LIBINDY_CRED_OFFER;
 use crate::utils::mockdata::mock_settings::StatusCodeMock;
-use crate::utils::parse_and_validate;
+use crate::indy::utils::parse_and_validate;
 use crate::{utils, WalletHandle};
 
 pub async fn libindy_issuer_create_credential_offer(
