+++ aries_vcx_core/src/indy/proofs/prover/mod.rs
@@ -6,7 +6,7 @@ use crate::errors::error::VcxCoreResult;
 use crate::global::settings;
 use crate::indy;
 use crate::utils::constants::REV_STATE_JSON;
-use crate::utils::parse_and_validate;
+use crate::indy::utils::parse_and_validate;
 
 pub async fn libindy_prover_create_revocation_state(
     tails_file_path: &str,
