+++ aries_vcx_core/src/indy/proofs/prover/prover.rs
@@ -5,7 +5,7 @@ use crate::errors::error::prelude::*;
 use crate::global::settings;
 use crate::indy::anoncreds::close_search_handle;
 use crate::utils::constants::{ATTRS, PROOF_REQUESTED_PREDICATES, REQUESTED_ATTRIBUTES};
-use crate::utils::parse_and_validate;
+use crate::indy::utils::parse_and_validate;
 use crate::{utils, WalletHandle};
 
 pub async fn libindy_prover_create_proof(
