+++ aries_vcx_core/src/indy/utils/mod.rs
@@ -1,5 +1,7 @@
 use std::sync::atomic::{AtomicUsize, Ordering};
-use vdrtools::CommandHandle;
+use vdrtools::{CommandHandle, types::validation::Validatable};
+
+use crate::errors::error::{VcxCoreResult, AriesVcxCoreError, AriesVcxCoreErrorKind};
 
 pub mod mocks;
 
@@ -7,6 +9,22 @@ static COMMAND_HANDLE_COUNTER: AtomicUsize = AtomicUsize::new(1);
 
 pub fn next_command_handle() -> CommandHandle {
     (COMMAND_HANDLE_COUNTER.fetch_add(1, Ordering::SeqCst) + 1) as CommandHandle
+}
+
+pub fn parse_and_validate<'a, T>(s: &'a str) -> VcxCoreResult<T>
+where
+    T: Validatable,
+    T: serde::Deserialize<'a>,
+{
+    let data = serde_json::from_str::<T>(s)?;
+
+    match data.validate() {
+        Ok(_) => Ok(data),
+        Err(s) => Err(AriesVcxCoreError::from_msg(
+            AriesVcxCoreErrorKind::LibindyInvalidStructure,
+            s,
+        )),
+    }
 }
 
 // TODO:  move to devsetup, see if we can reuse this / merge with different setup
