+++ aries_vcx_core/src/lib.rs
@@ -30,19 +30,13 @@ pub mod anoncreds;
 mod common;
 pub mod errors;
 pub mod global;
+#[cfg(feature = "vdrtools")]
 pub mod indy;
 pub mod ledger;
 pub mod utils;
 pub mod wallet;
 
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub struct WalletHandle(pub vdrtools::WalletHandle);
-pub const INVALID_WALLET_HANDLE: WalletHandle = WalletHandle(vdrtools::INVALID_WALLET_HANDLE);
-
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub struct SearchHandle(pub vdrtools::SearchHandle);
-pub const INVALID_SEARCH_HANDLE: SearchHandle = SearchHandle(vdrtools::INVALID_SEARCH_HANDLE);
-
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub struct PoolHandle(pub vdrtools::PoolHandle);
-pub const INVALID_POOL_HANDLE: PoolHandle = PoolHandle(vdrtools::INVALID_POOL_HANDLE);
+#[cfg(feature = "vdrtools")]
+pub use indy::{
+    PoolHandle, SearchHandle, WalletHandle, INVALID_POOL_HANDLE, INVALID_SEARCH_HANDLE, INVALID_WALLET_HANDLE,
+};
